requirements.txt
playwright==1.40.0
pytest==7.4.3
pytest-playwright==0.4.2
pytest-xdist==3.3.1
allure-pytest==2.13.2
pytest-html==4.1.1
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.5.0
requests==2.32.4 